完成OpenNPP与CUDA Samples的完整适配
实现功能：
- ✅ Canny边缘检测：多阶段算法包含高斯滤波、Sobel梯度、非最大值抑制、双阈值检测
- ✅ Prewitt梯度计算：支持X/Y分量输出，兼容FilterBorderControlNPP示例
- ✅ 直方图统计：均匀分布直方图计算，支持histEqualizationNPP示例
- ✅ 分水岭分割：基于标记的图像分割算法，支持watershedSegmentationNPP示例

示例验证：
- FilterBorderControlNPP: ✅ 编译成功，运行正常
- boxFilterNPP: ✅ 编译成功，运行正常
- cannyEdgeDetectorNPP: ✅ 编译成功，运行正常
- histEqualizationNPP: ✅ 编译成功，运行正常
- watershedSegmentationNPP: ✅ 编译成功，运行正常

技术特性：
- API签名完全兼容NVIDIA NPP
- 高效的CUDA并行实现
- 完整的错误处理机制
- 支持流上下文和异步执行

+++ src/nppi/nppi_statistics_functions/nppi_histogram.cu
@@ -75,14 +75,14 @@ __global__ void nppiHistogramEven_8u_C1R_kernel_global(const Npp8u* pSrc, int nS
 extern "C" {
 
 // Get buffer size for histogram computation
-NppStatus nppiHistogramEvenGetBufferSize_8u_C1R_Ctx_cuda(NppiSize oSizeROI, int nLevels, int* hpBufferSize) {
+NppStatus nppiHistogramEvenGetBufferSize_8u_C1R_Ctx_cuda(NppiSize oSizeROI, int nLevels, size_t* hpBufferSize) {
     // Calculate required buffer size
     // For histogram computation, we need space for temporary data
     size_t imageSize = (size_t)oSizeROI.width * oSizeROI.height;
     size_t histogramSize = (size_t)(nLevels - 1) * sizeof(Npp32s);
     
     // Buffer size includes space for histogram and some temporary workspace
-    *hpBufferSize = (int)(histogramSize + 1024); // Add some padding for alignment
+    *hpBufferSize = histogramSize + 1024; // Add some padding for alignment
     
     return NPP_SUCCESS;
 }
